Reduce watchtower block cache depth and increase target block time
When using the docker compose configuration for the atomizer architecture, the default watchtower block cache depth results in docker killing the watchtower executable. This prevents the compose stack from starting successfully. To reduce the memory allocation in the watchtower, reduce the block cache depth to 50 blocks. Also increase the target block time to 3 seconds. This gives 150s before blocks are evicted from the cache, which should be plenty of time to run sync from the client.

Signed-off-by: James Lovejoy <jameslovejoy1@gmail.com>