Fixed listen() "failing" to bind to invalid endpnt
Created a new m_archiver_invalid_endpoint variable to be able to add the editted variable
that has the invalid endpoint on it. Have to create new m_archiver_invalid_endpoint
variable because editting the vector that includes the invalid endpoint does not work since
a copy of the vector gets passed down according to controller initializer, not a reference of it.

The archiver test "archiver_server_init_failure" failed beforehand, which made it seem
as if connecting to an invalid endpoint still allowed the server to be initialized. Turns
out that the test was the problem. The test used to edit m_config_opts, however m_config_opts
did not get changed inside m_archiver, since m_config_opts was copied into m_archiver during
initialization, not referenced to. Now the test has a new m_archiver that has the m_config_opts
variable changes ever since initialization, allowing test to fail to bind to it.

closes #11

Signed-off-by: Enriko Chavez <enriko@mit.edu>