Make end-to-end atomizer test check the watchtower is syncing with the atomizer
If the watchtower fails to sync blocks from the atomizer, this was previously a silent failure. The tests would still fail but due to confusing side effects. This commit updates the test to check the watchtower is syncing blocks and that the mint transactions confirm. It also adds some useful log messages to components that were useful while investigating this issue.

Signed-off-by: James Lovejoy <jameslovejoy1@gmail.com>