Add build-folder flag
The previous implementation of test.sh assumed that the build folder was
set by an environment variable or that it was called 'build' and was located in
the top folder of the repo (i.e 'opencbdc-tx/build').  It also required
that the script be called from the top folder.  This commit adds a flag
that allows the build folder to be set via the command line.  It also adds the
ability for users to run the script from any folder.

Signed-off-by: Michael L. Szulczewski <mszulczewski@mitre.org>