Allow tcp_client::init to fail correctly
sentinel_2pc::controller::init contains a bug that makes it impossible to
trigger the error "Failed to start coordinator client" (see GitHub Issue #131).
This commit fixes the bug using the method described in Issue #131.  It
adds a new unit test to test triggering the error, and also modifies a
different unit test to account for behavior that's expected after the bug fix.

Signed-off-by: Michael L. Szulczewski <mszulczewski@mitre.org>