Fix bug that could cause non-existent element access
Previously, sentinel_2pc::controller would incorrectly calculate the
upper bound of the sentinel ID distribution if the number of sentinel
clients was 0.  The incorrect calculation was due to casting a negative
number to an unsigned int.  As a result of this error, the sentinel IDs
taken from the distribution could exceed the range of the possible
sentinel IDs, which could lead to a non-existent element access when a
bad ID was used as an index to a vector.

Signed-off-by: Michael L. Szulczewski <mszulczewski@mitre.org>