Check precondition when initializing sentinel 2PC controller
sentinel_2pc::controller allows the sentinel ID to exceed the number of
sentinels, which causes a non-existent element access (see GitHub
Issue #140).  This commit checks for that condition and throws an error if
it's true.  It also adds a unit test to confirm that the code behaves
correctly after the fix.

Signed-off-by: Michael L. Szulczewski <mszulczewski@mitre.org>