Check preconditions for initializing a locking-shard controller
Previously, the initialization of a locking shard controller could fail
in ways that were difficult to understand from the output to stdout.
This commit adds checks for things such as null pointers and out-of-range
vector accesses, and logs human-readable messages when the checks fail.
It also adds unit tests to ensure the checks work.

Signed-off-by: Michael L. Szulczewski <mszulczewski@mitre.org>