Refactor mock_system class for clarity
The mock_system class has a few parts that could cause confusion.
For example, the words 'module' and 'component' are used interchangeably,
it's not clear when modules will be mocked, and it's not clear what the
return value of start_servers means (see Issue #183).  This commit
refactors the code to clarify these issues.

Signed-off-by: Michael L. Szulczewski <mszulczewski@mitre.org>