Initial commit of Helm charts and e2e testing
This PR aims to improve validation of changes to opencbdc by setting up a Kubernetes cluster with minikube in Github actions and then install the 2PC architecture. After installation into the minikube cluster there are tests run to check whether or not pods/containers have started and are running. This is a very minimal test that currently does not check anything besides pod/container status. While we are using the Helm chart to run a test this exact configuration can be used to run in any Kubernetes cluster/environment. Documentation should be improved in a future PR to make it clear how to use the Helm chart and customize configuration.

The following has been added:
- Helm Chart for 2PC architecture to simplify running 2PC in Kubernetes
- Go tests using testify and Terratest to install the 2PC Helm Chart and check if containers are running

Signed-off-by: Kyle Crawshaw <kyle.crawshaw@gmail.com>