Abort if agent doesn't hold tickets on update keys
Addresses a bug identified which allows for a smart contract
to return an update map containing keys that the agent has
not acquired tickets for. This fix ensures that when smart
contract execution completes, all updates will be committed
or none of them will. Previously, committed changes which
the agent held write locks for would be committed, but changes
which never had ticket requests made would silently fail.
Note: if a key has a ticket but is waiting on a lock, the
transaction rolls back and retries. If a ticket has not
been acquired, the transaction should fail as the contract
specifically needs to request locks on important keys.

Signed-off-by: Michael Maurer <maurer.mi@northeastern.edu>