docs: mention new GH-Pages deployment in README
Before merging #255, the gh-pages branch of this repository was used
for our live API reference instance. Unfortunately, that branch houses
multiple binaries which were rebuilt on every-push, leading to the
branch growing in-size without bound, and dramatically increasing the
storage and network burden for all clones and forks.

Following #255 and the deletion of the gh-pages branch, the repo can
be cloned almost instantly, and the burden is fully offloaded to a
second repository which no one should need to clone under normal
circumstances.

This commit simply adds reference to the new deployment (since GH's
web interface is not smart enough to link to the external repo
automatically).

Signed-off-by: Sam Stuewe <stuewe@mit.edu>