feat: integrate proofs internally
This is the most invasive patch in the branch, and breaks a lot of
core assumptions (and, by extension) ancilliary tooling.

Practically speaking, the CI definitely fails at this commit (due to
the test suite not being updated at the same time). This patch is
split out because it hopefully contains all the core changes
specifically related to the source changes for the integration of
the new data structures (for easier review and reading); but it
should not (/ cannot) be targeted on its own.

Co-authorship (listed lexicographically) indicates either instrumental
troubleshooting, inspiring core aspects of the implementation, or
otherwise contributing significant aid, but is not intended to share
blame. If anything is wrong or broken in this patchset, blame lands at
the primary author's feet!

Co-authored-by: Alexander Jung <104335080+AlexRamRam@users.noreply.github.com>
Co-authored-by: Gert-Jaap Glasbergen <gertjaap@wadagso.com>
Co-authored-by: James Lovejoy <jameslovejoy1@gmail.com>
Signed-off-by: Sam Stuewe <stuewe@mit.edu>