fix: reconcile tests and benchmarks for proofs
This commit represents the updating of the rest of the testing and
benchmarking code to function as-expected with the new transaction
format and protocol. From this point forward, you should expect CI to
pass.

This is an omnibus update and includes help from multiple authors, I
have attempted to capture contributors below as co-authors.

Co-authored-by: Alexander Jung <104335080+AlexRamRam@users.noreply.github.com>
Co-authored-by: Gert-Jaap Glasbergen <gertjaap@wadagso.com>
Co-authored-by: James Lovejoy <jameslovejoy1@gmail.com>
Co-authored-by: Michael Maurer <maurer.mi@northeastern.edu>
Signed-off-by: Sam Stuewe <stuewe@mit.edu>